+++ libvast/include/vast/detail/env.hpp
@@ -8,7 +8,9 @@
 
 #pragma once
 
-#include <caf/expected.hpp>
+#include "vast/detail/generator.hpp"
+
+#include <caf/error.hpp>
 
 #include <optional>
 #include <string>
@@ -18,11 +20,26 @@ namespace vast::detail {
 /// A thread-safe wrapper around `::getenv`.
 /// @param var The environment variable.
 /// @returns The copied environment variables contents, or `std::nullopt`.
-[[nodiscard]] std::optional<std::string> locked_getenv(const char* var);
+[[nodiscard]] std::optional<std::string_view> getenv(std::string_view var);
+
+/// A thread-safe wrapper around `::setenv`.
+/// @param key The environment variable key.
+/// @param value The environment variable value.
+/// @param overwrite Flag to control whether existing keys get overwritten.
+/// @returns The copied environment variables contents, or `std::nullopt`.
+[[nodiscard]] caf::error
+setenv(std::string_view key, std::string_view value, int overwrite = 1);
 
 /// A thread-safe wrapper around `::unsetenv`.
 /// @param var The environment variable.
 /// @returns True on success.
-[[nodiscard]] caf::expected<void> locked_unsetenv(const char* var);
+[[nodiscard]] caf::error unsetenv(std::string_view var);
+
+/// Retrieves all environment variables as list of key-value pairs.
+/// The function processes the global variable `environ` that holds an array of
+/// strings, each of which has the form `key=value` by convention (per `man
+/// environ`). The results does not include variables that violate this
+/// convention.
+generator<std::pair<std::string_view, std::string_view>> environment();
 
 } // namespace vast::detail